(Context) Loading and FeaturedRooms components
The Context API was implemented in the context.js which allows informaton to be shared between a tree of React components.
The Loading and FeaturedRooms components were implemented.
Loading is a simple component displaying a gif while the data is being fetched from the server.
The FeaturedRooms component displays the featured rooms on the home screen